chore: reorganise visor commands (#445)
