+++ commands/walk.go
@@ -3,9 +3,14 @@ package commands
 import (
 	"context"
 	"errors"
+	"fmt"
+	"os"
 	"strings"
 	"time"
 
+	"github.com/filecoin-project/lotus/chain/actors/builtin"
+	lotuscli "github.com/filecoin-project/lotus/cli"
+	"github.com/filecoin-project/sentinel-visor/lens/lily"
 	"github.com/filecoin-project/sentinel-visor/schedule"
 	"github.com/urfave/cli/v2"
 	"golang.org/x/xerrors"
@@ -15,7 +20,79 @@ import (
 	"github.com/filecoin-project/sentinel-visor/storage"
 )
 
-var Walk = &cli.Command{
+type walkOps struct {
+	from   int64
+	to     int64
+	tasks  string
+	window time.Duration
+}
+
+var walkFlags walkOps
+
+var WalkCmd = &cli.Command{
+	Name:   "walk",
+	Usage:  "Start a daemon job to walk a range of the filecoin blockchain.",
+	Before: initialize,
+	After:  destroy,
+	Flags: []cli.Flag{
+		&cli.StringFlag{
+			Name:        "tasks",
+			Usage:       "Comma separated list of tasks to run. Each task is reported separately in the database.",
+			Value:       strings.Join([]string{chain.BlocksTask, chain.MessagesTask, chain.ChainEconomicsTask, chain.ActorStatesRawTask}, ","),
+			Destination: &walkFlags.tasks,
+		},
+		&cli.DurationFlag{
+			Name:        "window",
+			Usage:       "Duration after which any indexing work not completed will be marked incomplete",
+			Value:       builtin.EpochDurationSeconds * time.Second * 10, // walks don't need to complete within a single epoch
+			Destination: &walkFlags.window,
+		},
+		&cli.Int64Flag{
+			Name:        "from",
+			Usage:       "Limit actor and message processing to tipsets at or above `HEIGHT`",
+			Destination: &walkFlags.from,
+		},
+		&cli.Int64Flag{
+			Name:        "to",
+			Usage:       "Limit actor and message processing to tipsets at or below `HEIGHT`",
+			Value:       estimateCurrentEpoch(),
+			DefaultText: "MaxInt64",
+			Destination: &walkFlags.to,
+		},
+	},
+	Action: func(cctx *cli.Context) error {
+		ctx := lotuscli.ReqContext(cctx)
+
+		cfg := &lily.LilyWalkConfig{
+			Name:                "lily",
+			Tasks:               strings.Split(walkFlags.tasks, ","),
+			Window:              walkFlags.window,
+			From:                walkFlags.from,
+			To:                  walkFlags.to,
+			RestartDelay:        0,
+			RestartOnCompletion: false,
+			RestartOnFailure:    false,
+			Database: &lily.LilyDatabaseConfig{
+				URL:                  VisorCmdFlags.DB,
+				Name:                 VisorCmdFlags.Name,
+				PoolSize:             VisorCmdFlags.DBPoolSize,
+				AllowUpsert:          VisorCmdFlags.DBAllowUpsert,
+				AllowSchemaMigration: VisorCmdFlags.DBAllowMigrations,
+			},
+		}
+
+		watchID, err := lilyAPI.LilyWalk(ctx, cfg)
+		if err != nil {
+			return err
+		}
+		if _, err := fmt.Fprintf(os.Stdout, "Created Watch Job: %d", watchID); err != nil {
+			return err
+		}
+		return nil
+	},
+}
+
+var RunWalkCmd = &cli.Command{
 	Name:  "walk",
 	Usage: "Walk a range of the filecoin blockchain and process blocks as they are discovered.",
 	Flags: []cli.Flag{
@@ -43,78 +120,76 @@ var Walk = &cli.Command{
 			Hidden: true,
 		},
 	},
-	Action: walk,
-}
+	Action: func(cctx *cli.Context) error {
+		// Validate flags
+		heightFrom := cctx.Int64("from")
+		heightTo := cctx.Int64("to")
 
-func walk(cctx *cli.Context) error {
-	// Validate flags
-	heightFrom := cctx.Int64("from")
-	heightTo := cctx.Int64("to")
+		if heightFrom > heightTo {
+			return xerrors.Errorf("--from must not be greater than --to")
+		}
 
-	if heightFrom > heightTo {
-		return xerrors.Errorf("--from must not be greater than --to")
-	}
+		tasks := strings.Split(cctx.String("tasks"), ",")
 
-	tasks := strings.Split(cctx.String("tasks"), ",")
+		if err := setupLogging(cctx); err != nil {
+			return xerrors.Errorf("setup logging: %w", err)
+		}
 
-	if err := setupLogging(cctx); err != nil {
-		return xerrors.Errorf("setup logging: %w", err)
-	}
+		if err := setupMetrics(cctx); err != nil {
+			return xerrors.Errorf("setup metrics: %w", err)
+		}
 
-	if err := setupMetrics(cctx); err != nil {
-		return xerrors.Errorf("setup metrics: %w", err)
-	}
-
-	tcloser, err := setupTracing(cctx)
-	if err != nil {
-		return xerrors.Errorf("setup tracing: %w", err)
-	}
-	defer tcloser()
-
-	lensOpener, lensCloser, err := setupLens(cctx)
-	if err != nil {
-		return xerrors.Errorf("setup lens: %w", err)
-	}
-	defer func() {
-		lensCloser()
-	}()
-
-	var strg model.Storage = &storage.NullStorage{}
-	if cctx.String("csv") != "" {
-		csvStorage, err := storage.NewCSVStorage(cctx.String("csv"))
+		tcloser, err := setupTracing(cctx)
 		if err != nil {
-			return xerrors.Errorf("new csv storage: %w", err)
+			return xerrors.Errorf("setup tracing: %w", err)
 		}
-		strg = csvStorage
-	} else {
-		if cctx.String("db") == "" {
-			log.Warnw("database not specified, data will not be persisted")
-		} else {
-			db, err := setupDatabase(cctx)
+		defer tcloser()
+
+		lensOpener, lensCloser, err := setupLens(cctx)
+		if err != nil {
+			return xerrors.Errorf("setup lens: %w", err)
+		}
+		defer func() {
+			lensCloser()
+		}()
+
+		var strg model.Storage = &storage.NullStorage{}
+		if cctx.String("csv") != "" {
+			csvStorage, err := storage.NewCSVStorage(cctx.String("csv"))
 			if err != nil {
-				return xerrors.Errorf("setup database: %w", err)
+				return xerrors.Errorf("new csv storage: %w", err)
 			}
-			strg = db
+			strg = csvStorage
+		} else {
+			if cctx.String("db") == "" {
+				log.Warnw("database not specified, data will not be persisted")
+			} else {
+				db, err := setupDatabase(cctx)
+				if err != nil {
+					return xerrors.Errorf("setup database: %w", err)
+				}
+				strg = db
+			}
 		}
-	}
 
-	tsIndexer, err := chain.NewTipSetIndexer(lensOpener, strg, 0, cctx.String("name"), tasks)
-	if err != nil {
-		return xerrors.Errorf("setup indexer: %w", err)
-	}
+		tsIndexer, err := chain.NewTipSetIndexer(lensOpener, strg, 0, cctx.String("name"), tasks)
+		if err != nil {
+			return xerrors.Errorf("setup indexer: %w", err)
+		}
 
-	scheduler := schedule.NewScheduler(cctx.Duration("task-delay"),
-		&schedule.JobConfig{
-			Name:                "Walker",
-			Job:                 chain.NewWalker(tsIndexer, lensOpener, heightFrom, heightTo),
-			RestartOnFailure:    false, // Don't restart after a failure otherwise the walk will start from the beginning again
-			RestartOnCompletion: false,
-			RestartDelay:        time.Minute,
-		})
+		scheduler := schedule.NewScheduler(cctx.Duration("task-delay"),
+			&schedule.JobConfig{
+				Name:                "Walker",
+				Job:                 chain.NewWalker(tsIndexer, lensOpener, heightFrom, heightTo),
+				RestartOnFailure:    false, // Don't restart after a failure otherwise the walk will start from the beginning again
+				RestartOnCompletion: false,
+				RestartDelay:        time.Minute,
+			})
 
-	err = scheduler.Run(cctx.Context)
-	if !errors.Is(err, context.Canceled) {
-		return err
-	}
-	return nil
+		err = scheduler.Run(cctx.Context)
+		if !errors.Is(err, context.Canceled) {
+			return err
+		}
+		return nil
+	},
 }
